refactor codebase to separate culture, decision, and general purpose utilities
